ai-guardrail-webhook-server/pyproject.toml
[project]
name = "ai-guardrail-webhook-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "opentelemetry-exporter-otlp-proto-grpc>=1.31.1",
    "opentelemetry-instrumentation-fastapi>=0.52b1",
    "pydantic-core>=2.33.1",
    "pyyaml>=6.0.2",
    "uvicorn>=0.34.0",
]
